Add cpanato as reviewer
In an effort to decrease the bus factor for slack-infra, this commit
adds cpanato as a reviewer.

Carlos has written the `slack-moderator-words` bot and is familiar with
the code. Additionally, he's also been involved in driving the bot's
deployment so he's familiar with the k8s.io side for slack-infra too.